Changes (PR) | moved model to separate file , added viewModel class
