+++ composeApp/src/commonMain/kotlin/dev/androidpoet/weatherapp/data/remote/WeekForeCastComposeModel.kt
@@ -1,0 +1,47 @@
+package dev.androidpoet.weatherapp.data.remote
+
+data class WeekForeCastComposeModel(
+    val dt: Long,
+    val sunrise: Long,
+    val sunset: Long,
+    val moonrise: Long,
+    val moonset: Long,
+    val moonPhase: Double,
+    val summary: String,
+    val temp: Temp,
+    val feelsLike: FeelsLike,
+    val pressure: Long,
+    val humidity: Long,
+    val dewPoint: Double,
+    val windSpeed: Double,
+    val windDeg: Long,
+    val windGust: Double,
+    val weather: List<Weather>,
+    val clouds: Long,
+    val pop: Double,
+    val rain: Double,
+    val uvi: Double,
+)
+
+data class Temp(
+    val day: Double,
+    val min: Double,
+    val max: Double,
+    val night: Double,
+    val eve: Double,
+    val morn: Double,
+)
+
+data class FeelsLike(
+    val day: Double,
+    val night: Double,
+    val eve: Double,
+    val morn: Double,
+)
+
+data class Weather(
+    val id: Long,
+    val main: String,
+    val description: String,
+    val icon: String,
+)
