+++ composeApp/src/commonMain/kotlin/dev/androidpoet/weatherapp/presentation/viewmodel/WeatherForeCastViewModel.kt
@@ -1,0 +1,4 @@
+package dev.androidpoet.weatherapp.presentation.viewmodel
+
+class WeatherForeCastViewModel {
+}
